+++ node/runtime/pangolin/src/pallets/mod.rs
@@ -121,8 +121,8 @@ pub use relay_authorities::*;
 pub mod to_tron_backing_;
 pub use to_tron_backing_::*;
 
-pub mod evm_;
-pub use evm_::*;
+pub mod evm;
+pub use evm::*;
 
 pub mod ethereum;
 pub use ethereum::*;
